Rename SSSOM source
